requirements.txt
poetry
google-adk[cli]
python-dotenv
pandas
pyarrow